requirements.txt
aiogram==3.7.0
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.2
ruff==0.1.8